+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This CHANGELOG details important changes made in each version of the
 
 ## v0.18.3 (Unreleased)
 
+### Improvements
+
+- Automatically generate `isInstance` type guards for implementations of `Resource`.
+
 ## v0.18.2 (Released May 28th, 2019)
 
 - Improved the package `README` file to reflect usage of the `@pulumi/terraform`